% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bgs}
\alias{bgs}
\title{Moustiques adultes}
\format{
\code{data.frame}
}
\source{
Projet OpTIS
}
\usage{
bgs
}
\description{
Table des captures de moustiques adultes échantillonnés avec des pièges BG Sentinel.
}
\keyword{datasets}
