% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansd.R
\name{cleansd.lvlbg}
\alias{cleansd.lvlbg}
\title{Nettoyage des données BGS de suppression}
\usage{
\method{cleansd}{lvlbg}(data, treated, end = Sys.Date(), ...)
}
\arguments{
\item{data}{objet de classe \code{lvlbg} output de la fonctions
d'importation \code{getBGSdata}}

\item{treated}{scalaire de type caractère indiquant le site
traité: \code{bsit1} ou \code{bsit2}.}

\item{end}{scalaire de type date indiquant la date limite
supérieure pour laquelle on veut les taux de
suppression. Valeur par défaut \code{Sys.Date()}.}

\item{...}{ellipse: argument(s) supplémentaires passés aux
méthodes spécifiques (aucun pour le moment).}
}
\value{
un objet de classe \code{cleaned} héritant de la classe
\code{data.frame}, avec 4 champs:
\itemize{
\item \code{treatment} facteur: site du piège;
\item \code{piege_id} facteur: identifiant du piège,
\item \code{period} facteur: période de capture,
\item \code{m} entier: densité apparente.
}
}
\description{
Méthode de nettoyage des données BGS nécessaires
pour estimer la suppression des femelles adultes \emph{Aedes
albopictus}.
}
\keyword{cleaning}
