% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansd.R
\name{cleansd.lvlov}
\alias{cleansd.lvlov}
\title{{Nettoyage des données ovitrap de suppression}}
\usage{
\method{cleansd}{lvlov}(data, treated, end = Sys.Date(), ...)
}
\arguments{
\item{data}{objet de classe \code{lvlov} output de la fonctions
d'importation \code{getOVIdata}}

\item{treated}{scalaire de type caractère indiquant le site
traité: \code{bsit1} ou \code{bsit2}.}

\item{end}{scalaire de type date indiquant la date limite
supérieure pour laquelle on veut les taux de
suppression. Valeur par défaut \code{Sys.Date()}.}

\item{...}{ellipse: argument(s) supplémentaires passés aux
méthodes spécifiques (aucun pour le moment).}
}
\value{
un objet de classe \code{cleaned} héritant de la classe
\code{data.frame}, avec 4 champs: * \code{treatment} facteur: site du
piège; * \code{piege_id} facteur: identifiant du piège, * \code{period}
facteur: période de capture, * \code{m} entier: densité apparente.
}
\description{
Méthode de nettoyage des données ovitrap nécessaires
pour estimer la suppression des oeufs.
}
\keyword{data}
\keyword{management}
