% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleansd.R
\name{cleansd}
\alias{cleansd}
\title{{Nettoyage des données de suppression}}
\usage{
cleansd(data, treated, end, ...)
}
\arguments{
\item{data}{objet de classe \code{lvlov} ou \code{lvlbg}; héritant
de la classe \code{data.frame}, output des fonctions
d'importation \code{getOVIdata} - échantillonnage des oeufs
d'\emph{Aedes}, ou \code{getBGSdata} - échantillonnage des
moustiques adultes.}

\item{treated}{scalaire de type caractère indiquant le site
traité: \code{bsit1} ou \code{bsit2}.}

\item{end}{scalaire de type date indiquant la date limite
supérieure pour laquelle on veut les taux de
suppression.}

\item{...}{ellipse: argument(s) supplémentaires passés aux
méthodes spécifiques (aucun pour le moment).}
}
\value{
\code{data.frame}
}
\description{
Méthode générique de nettoyage des données de
suppression
}
\keyword{cleaning}
