% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emprise}
\alias{emprise}
\title{Emprise du projet OpTIS}
\format{
Objet de classe \code{RasterLayer} (package raster), projection EPSG:2975. à 3 bandes (R, G, B).
}
\source{
CartoDB.PositronNoLabels
}
\usage{
emprise
}
\description{
Fond de carte utilisé dans les cartes de densité relative des moustiques.
}
\keyword{datasets}
