% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBGSdata.R
\name{getBGSdata}
\alias{getBGSdata}
\title{{Importe les captures de moustiques adultes}}
\usage{
getBGSdata()
}
\value{
{un data.frame}
}
\description{
{Cette fonction importe depuis le système
d'information Laveval les résultats des comptages de
moustiques adultes capturés avec des pièges BG Sentinel
appâtés au CO~2~ gazeux, pendant le monitoring entomologique
du projet OpTIS.}
}
\details{
{Les données de trois espèces sont disponibles dans
Laveval: (i) \emph{Aedes albopictus}, (ii)\emph{Aedes argypti} et
(iii) \emph{Culex quinquefasciattus}.
Pour chaque enregistrement, on a (liste non exhaustive):
\itemize{
\item le site concerné,
\item la zone Alizé concernée,
\item l'identifiant du piàge,
\item la date de pose du piège,
\item la date de relevé du piège,
\item le nombre de mâles sauvages \emph{Aedes albopictus},
\item le nombre de mâles stériles_Aedes albopictus_
\item le nombre de femelles sauvages \emph{Aedes albopictus},
\item le nombre de femelles stériles \emph{Aedes albopictus},
\item le nombre d'adultes \emph{Aedes albopictus} de sexe indéterminé,
\item le nombre de mâles stériles_Aedes albopictus_ jaunes,
\item le nombre de mâles stériles_Aedes albopictus_ orange,
\item le nombre de mâles stériles_Aedes albopictus_ rose,
\item le nombre de mâles stériles_Aedes albopictus_ bleu,
\item le nombre de mâles stériles_Aedes albopictus_ vert,
\item le nombre de mâles sauvages \emph{Aedes aegypti},
\item le nombre de femelles sauvages \emph{Aedes  aegypti},
\item le nombre d'adultes \emph{Aedes  aegypti} de sexe indéterminé,
\item le nombre de mâles sauvages \emph{Culex quinquefasciattus},
\item le nombre de femelles sauvages \emph{Culex quinquefasciattus},
\item le nombre d'adultes \emph{Culex quinquefasciattus} de sexe indéterminé}
}
}
\keyword{import}
