% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBkg.R
\name{getBkg}
\alias{getBkg}
\title{{Fond sz carte de la zone d'emprise du projet OpTIS}}
\usage{
getBkg(pos = c(0.75, 0.05), scalem)
}
\arguments{
\item{pos}{vecteur numérique de longueur 2 donnant la position (x,
y) en coordonnées (0, 1) de l'extrémité gauche de l'échelle de
l'emprise.}

\item{scalem}{scalaire numérique donnant la longueur de la barre
d'échelle, en mètres}
}
\value{
{un graphique de classe \code{lattice}}
}
\description{
Cette fonction produit un fond de carte de la zone
d'emprise du projet OpTIS, sous la forme d'un graphique
produit par la fobctiob \code{rasterVis::levelplot}. Ce
graphique est utilisable comme fond de xarte dans les
graphiques gu paxkage recommandé \code{lattice} grâce à la
fonction \code{latticeExtra::as.layer}
}
\keyword{graphics}
