% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCON.R
\name{getCON}
\alias{getCON}
\title{{Connexion à LAVeval}}
\usage{
getCON(env = "LAVEVAL_DB_PASSWORD")
}
\arguments{
\item{env}{{nom de la variable d'environnement contenant le mot de
passe pour la base de données Laveval}}
}
\value{
{une connexion est ouverte}
}
\description{
{Cette fonction ouvre une connexion à la base de
données Laveval}
}
\details{
{Cette fonction n'a pas vocation à être utiisée directement. Elle est urilisée par les fonctions d'importation du package.}
}
\keyword{internal}
