% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOVIdata.R
\name{getOVIdata}
\alias{getOVIdata}
\title{{Importe les comptages d'oeufs d'\emph{Aedes}}}
\usage{
getOVIdata()
}
\value{
{un data.frame avec tout c kil fô dedans ;-}
}
\description{
{Cette fonction importe depuis le système
d'information Laveval les comptages d'oeufs d'\emph{Aedes}
collectés avec ds ovitraps pendant le monitoring entomologique
du projet OpTIS.}
}
\keyword{import}
