% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTraps.R
\name{getTraps}
\alias{getTraps}
\title{{Importe la table des pièges}}
\usage{
getTraps()
}
\value{
un \code{data.frame}
}
\description{
Cette fonction importe la table de tous les pièges enregistrés dans Laveval
}
\keyword{import}
