% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getZA.R
\name{getZA}
\alias{getZA}
\title{{Importe les zones Alizé}}
\usage{
getZA()
}
\value{
collection de polygones spatiaux sf
}
\description{
{Cette fonction importe la tabke dee polygons des zones Alizé}
}
\keyword{import}
