% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovi}
\alias{ovi}
\title{Oeufs d'\emph{Aedes}}
\format{
\code{data.frame}
}
\source{
Projet OpTIS
}
\usage{
ovi
}
\description{
Table des oeufs d'\emph{Aedes}  échantillonnés avec des ovitraps
}
\keyword{datasets}
