% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{run}
\alias{run}
\title{Ligne de côte de La Réunion}
\format{
Objet de classe \code{SpatialPolygons} (package sp), projection EPSG:2975.:
}
\source{
\url{https://data.nextgis.com/fr/region/RE/base/}
}
\usage{
run
}
\description{
Objet spatial (package \code{sp}) représentant la ligne de côte de La
Réunion. Source: La Réunion: fond de carte OpenStreetMap.
}
\keyword{datasets}
