% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{traps}
\alias{traps}
\title{Pièges}
\format{
Objet de classes \code{sf} et \code{data.frame}: collection de points spatiaux..
}
\source{
Projet OpTIS
}
\usage{
traps
}
\description{
Table des pièges (BGS et ovitraps) utilisés pour le monitoring des populations de moustisues dans le projet OpTIS
}
\keyword{datasets}
