% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{za}
\alias{za}
\title{Zones Alisé}
\format{
Objet de classes \code{sf} et \code{data.frame}: collection de polygonrs spatiaux.
}
\source{
ARS La Réunion
}
\usage{
za
}
\description{
Collection de Polygones spatiaux des zones Alizé.
}
\keyword{datasets}
